requirements.txt
aiohttp==3.5.4
pyyaml==5.1.2
